+++ src/components/configurable-items/configurable-items.style.js
@@ -6,7 +6,7 @@ import {
   ConfigurableItemRowStyle,
 } from "./configurable-item-row/configurable-item-row.style";
 import StyledFormField from "../../__experimental__/components/form-field/form-field.style";
-import StyledCheckbox from "../../__experimental__/components/checkbox/checkbox.style";
+import StyledCheckbox from "../checkbox/checkbox.style";
 
 import baseTheme from "../../style/themes/base";
