+++ src/components/draggable/draggable.stories.mdx
@@ -2,7 +2,7 @@ import { Meta, Story, Preview } from "@storybook/addon-docs/blocks";
 
 import StyledSystemProps from "../../../.storybook/utils/styled-system-props";
 import { DraggableContainer, DraggableItem } from ".";
-import { Checkbox } from "../../__experimental__/components/checkbox";
+import { Checkbox } from "../checkbox";
 
 <Meta
   title="Design System/Draggable"
