+++ src/components/fieldset/fieldset.stories.js
@@ -2,7 +2,7 @@ import React from "react";
 import { text } from "@storybook/addon-knobs";
 import Fieldset from "./fieldset.component";
 import Textbox from "../textbox";
-import { Checkbox } from "../../__experimental__/components/checkbox";
+import { Checkbox } from "../checkbox";
 
 export default {
   title: "Design System/Fieldset/Test",
